Cleanup and updated documentation
