+++ doc/aes_128.md
@@ -1,14 +1,14 @@
 # Introduction
 AES is a symmetric-key encryption algorithm based on a design principle known as a substitution-permutation network. Refer to [FIPS-197](https://csrc.nist.gov/pubs/fips/197/final) for detailed specifications. This project implements AES-128 on an inexpensive IO-limited FPGA using Cipher Block Chaining (CBC).
 
-### Figure 1: AES CBC Mode (Encryption)
-<img src="figures/cbc_encryption.png" alt="" width="500"/>
+<!-- ### Figure 1: AES CBC Mode (Encryption)
+<img src="figures/cbc_encryption.png" alt="" width="250"/>
 
 ### Figure 2: AES CBC Mode (Decryption)
-<img src="figures/cbc_decryption.png" alt="" width="500"/>
+<img src="figures/cbc_decryption.png" alt="" width="250"/> -->
 
-### Figure 3: AES Encrypt Block Diagram
-<img src="figures/aes_cbc_bd.drawio.png" alt="" width="250"/>
+### Figure 1: AES Encrypt Block Diagram
+<img src="figures/aes_cbc_bd.drawio.png" alt="" width="500"/>
 
 # Components
 The design is broken up into the following components:
@@ -35,7 +35,11 @@ The design is broken up into the following components:
     - Inverse of shift_rows
 - inv_mix_columns
     - Inverse of mix_columns
-    
+
+These components are arranged in the following way, with additional logic where necessary:
+### Figure 2: Block Diagram
+<img src="figures/block_diagram.drawio.png" alt="" width="1000"/>
+
 # External Interface
 
 For detailed information on the external interface, refer to the [interface specification](external_interface.md).
@@ -53,4 +57,8 @@ For detailed information on the external interface, refer to the [interface spec
 
 To run testbenches, follow the [environment setup](env-setup.md). Debian on WSL was used for the setup instructions, but the basic steps should remain the same.
 
-Run `python3 {testname.py}` to run a test.
+Run `python3 {testname.py}` to run a test. `aes_128_top_wrapper_test.py` interfaces with `aes_128_top_wrapper.vhdl` through the external interface and implements three tests:
+
+1. Tests the DUT based on FIPS-197 Appendix B, with one round of encryption and decryption.
+2. Tests the DUT using random initial vector, key, and plaintext, with one round of encryption and decryption.
+3. Tests the CBC mode of the DUT, encrypting and decrypting a string of words. Checks the outputs against the same string encrypted with the "pycryptodome" python library.
