+++ src/enc/aes_128_top_enc.vhdl
@@ -1,38 +1,28 @@
 library ieee;
 use ieee.std_logic_1164.all;
-library work;
 use work.aes_pkg.all;
 
 entity aes_128_top_enc is
-generic
-(
-    IBW : natural range 1 to 16;  -- Input Bus Width must be power of 2
-    OBW : natural range 1 to 16   -- Output Bus Width must be power of 2
-);
 port 
 (
     -- Common
     clk               : in std_logic;
     reset             : in std_logic;
     -- Input
-    input_bus         : in std_logic_vector(IBW*8-1 downto 0);
+    input_bus         : in std_logic_vector(127 downto 0);
     e_key             : in exp_key_type;
     init_vec          : in std_logic_vector(127 downto 0); -- initial vector to XOR with the plaintext
     init_vec_valid    : in std_logic; 
     -- all inputs must remain valid whenever input_valid is pulsed, even if they're from previous rounds
     input_valid       : in std_logic;                      -- Inlcudes everything: bus, key, init_vec, session_start
     -- Output
-	cipherblock       : out std_logic_vector(OBW*8-1 downto 0);
+	cipherblock       : out std_logic_vector(127 downto 0);
     output_valid      : out std_logic
 );
 end aes_128_top_enc;
 
 architecture rtl of aes_128_top_enc is
-    constant IBW_ROUNDS              : natural := 16/IBW - 1; -- # of ccs required to read in the plaintext
-    constant OBW_DELAY               : natural := 16/OBW - 1; -- # of ccs required to output cipherblock
-    signal round_key_valid           : std_logic;
     signal plaintext                 : std_logic_vector(127 downto 0);
-    signal input_index               : natural range 0 to 16; 
     signal input_block_ready         : std_logic;
     
     signal prev_cipherblock          : std_logic_vector(127 downto 0);
@@ -48,60 +38,14 @@ architecture rtl of aes_128_top_enc is
     
     signal shift_rows_bus_out_valid  : std_logic;
     signal shift_rows_bus_out        : std_logic_vector(127 downto 0);
-        
-    signal key_ready                 : std_logic;
 
-    type key_proc_state_type is (idle, start_round, enc_in_prog, end_enc);
+    type key_proc_state_type is (idle, enc_in_prog, end_enc);
     signal rnd_key_state             : key_proc_state_type;
     signal xor_init_vec : std_logic;
     signal xor_init_vec_done : std_logic;
 begin
-    -- Process for buffering the input plaintext
-    cntrl_proc : process(clk)
-    variable t_index : integer;
-    variable b_index : integer;
-    begin
-        if rising_edge(clk) then
-            input_block_ready <= '0'; --Reset pulse
-            xor_init_vec_done <= '0';
-            if reset = '1' then
-                input_index <= 0;
-            else
-                if init_vec_valid = '1' then
-                    xor_init_vec <= '1';
-                elsif xor_init_vec_done = '1' then
-                    xor_init_vec <= '0';
-                end if;
-                if input_valid = '1' or input_index = IBW_ROUNDS + 1 then
-                    if input_index = IBW_ROUNDS + 1 then
-                        -- we've read in the entire plaintext
-                        if xor_init_vec = '1' then
-                            -- this will override the previous cipherblock
-                            plaintext <= plaintext xor init_vec; -- xor with initial vector
-                            input_index <= 0;
-                            input_block_ready <= '1'; -- Pulsed
-                            xor_init_vec_done <= '1'; -- Pulsed
-                        elsif prev_cipherblock_valid = '1' then
-                            -- xor with prev cipherblock
-                            plaintext <= plaintext xor prev_cipherblock;
-                            input_index <= 0;
-                            input_block_ready <= '1'; -- Pulsed
-                        end if;
-                    else
-                        -- more to go
-                        -- Variables for top and bottom bounds of assigning state array.
-                        t_index := 127 - input_index*IBW*8;
-                        b_index := 128 - (1 + input_index)*IBW*8;
-
-                        plaintext(t_index downto b_index) <= input_bus;
-                        input_index <= input_index + 1;
-                    end if;
-                end if; -- input valid
-            end if; -- reset
-        end if; -- clk
-    end process;
     
-    -- -- Process to add round key
+    -- Process to add round key
     add_round_key : process(clk)
         variable rnd_num : integer range 0 to 10;
         variable mix_col_out_rdy : std_logic;
@@ -109,17 +53,32 @@ begin
         if rising_edge(clk) then
             s_box_bus_in_valid <= '0'; -- Clear pulses
             output_valid <= '0';
+            input_block_ready <= '0';
+            xor_init_vec_done <= '0';
             if reset = '1' then
                 rnd_key_state <= idle;
                 rnd_num := 0;
                 mix_col_out_rdy := '0';
             else
-                if input_block_ready = '1' then
+                -- Control signals
+                if init_vec_valid = '1' then
+                    xor_init_vec <= '1';
+                elsif xor_init_vec_done = '1' then
+                    xor_init_vec <= '0';
+                end if;
+
+                if input_valid = '1' then
+                    if xor_init_vec = '1' then
+                        -- this will override the previous cipherblock
+                        s_box_bus_in <= input_bus xor init_vec xor e_key(0); -- xor with initial vector
+                        xor_init_vec_done <= '1'; -- Pulsed
+                    elsif prev_cipherblock_valid = '1' then
+                        -- xor with prev cipherblock
+                        s_box_bus_in <= input_bus xor prev_cipherblock xor e_key(0);
+                    end if;
                     -- This initiates the encryption round
-                    rnd_key_state <= start_round;
+                    input_block_ready <= '1'; -- Pulsed
                     prev_cipherblock_valid <= '0';
-                    -- In this case we xor with the state array
-                    s_box_bus_in <= plaintext xor e_key(0);
                     s_box_bus_in_valid <= '1'; -- Pulsed
                     rnd_key_state <= enc_in_prog;
                     rnd_num := 0;
