+++ src/enc/s_box.vhdl
@@ -1,6 +1,5 @@
 library ieee;
 use ieee.std_logic_1164.all;
-library work;
 use work.aes_pkg.all; -- s_box_byte
 
 entity s_box is
